Setup fancy statusbar on nightwing
I've just been tweaking the colours in my tmux config, so that they'll
look better with the Vim jellybeans theme. When I took a look at the
finished result I felt the standard black on white vim status bar was
rather letting the side down.

Debian being what it is, I thought it might be a piece of cake to setup
a powerline-style status bar and get it looking nice alongside my tmux
tweaks. And so it proved.

I chose to use the vim-airline plugin (rather than the powerline
project) as it came up first on Google, but have since discovered that
powerline [1] is likely to be more powerful. powerline was ported to
Python, which adds an extra dependency and I can imagine it wouldn't
have such good performance, so vim-airline (which just worked as soon as
I installed it) will do me very nicely for now.

Note that in addition to installing the plugin, to enable the
arrow-style characters in the status bar you need to install the
powerline fonts [1]. Then you must set a config variable to initialise
them.  I've only set that variable on nightwing, as it's currently the
only machine with the fonts installed.

By default the jellybeans airline plugin is used (to match the vim
theme), but I found the colours of the lucius theme a little less
distracting. So I've also configured nightwing to use the lucius theme.

[1] https://github.com/powerline
[2] https://github.com/powerline/fonts