Install jedi-vim for Python autocompletion
It uses Jedi to implement IDE-style autocompletion in Python code. It
also supports jumping to definitions, highlighting usage of something,
and renaming stuff. Very nifty.

By default it will pop up a list of all the contents of a module/object
as you type, which I think I'd find rather irritating (and very
different to the way I use Vim's built-in completion when writing in
other languages), so I've disabled the auto-popup feature in .vimrc.