ctrl-p lists files in current working directory
In large git repositories that contain multiple apps, even though I was
launching vim within a sub folder of the repo, all files in the git repo
were listed when using ctrl-p to find files.

This is because the `g:ctrlp_user_command` is set to treat .git as a
root of the project.

Setting `g:ctrlp_working_path_mode` to 'a' means that it'll only find
files within the working directory. See the docs for that variable for
more details. The value of `g:ctrlp_working_path_mode` wasn't set by
default.