Make format=flowed highlighting less overpowering
ErrorMsg is red in my current colour scheme, and highlighting them as
errors isn't sensible anyway; I don't want to become de-sensitised to
the intent of that highlighting.

I went through all the docs for the different highlight modes, and at
least in my current colourscheme, Visual seems to be one that is nicely
low key. It's also has the benefit of always going to highlight
whitespace somehow, as it's the highlighting that Vim uses to indicate a
selection in visual mode.