Enable black if .use-black exists
Touch the `.use-black` dotfile in a project in order to automatically
reformat its code with Black.

I've limited the `BufWritePre` command to work on code within my
`~/Projects` folder so that any installed packages whose code I might be
messing about with don't get reformatted on save. Not sure I really need
to worry about that, but I can imagine it could happen and would be a
right pain.