Enable base16 colour theme
To choose a scheme the Preview page is useful:

  http://chriskempson.com/projects/base16/

Note that the preview page doesn't include some of the themes from the
fnune fork that I'm using (see previous commit).

I'm starting out with the circus theme, which seems nice and colourful
and has enough contrast to work well on all my screens (I've still got a
couple of laptops with TN panels).