Update vim-airline plugin to v0.11
This change pulss in 236 commits since I last updated it. The master
branch has seen another 430 on top of that, but v0.11 is the latest
release. There appears to be a bunch of stuff for Neovim coming in, and
my themes looked messed up when I updated to master. So 0.11 is far
enough now.