Update ag plugin to latest (deprecated) version
It's seen a lot of commits since I last updated it, and (partly due to
licensing issues) has been deprecated. It seems it was originally forked
from the ack.vim plugin that I removed two days ago, which the author is
recommending people use with this config in .vimrc:

    let g:ackprg = 'ag --vimgrep --smart-case'
    cnoreabbrev ag Ack
    cnoreabbrev aG Ack
    cnoreabbrev Ag Ack
    cnoreabbrev AG Ack

As I have the vim-ripgrep plugin installed I think I'll just map the :Ag
command to :Rg to avoid breaking my muscle memory, and see how I get on.

I'll do that in another commit.