Install plugins as submodules for use with minpac
This commit re-installs the plugins that I removed in my previous
commit, cloning them afresh in a location that Vim 8 and Neovim will
load automatically.

I'm using `git submodule add` to install new plugins, and minpac's
PackUpdate command to pull in and review updates. I'll commit any
updates it pulls down as you would for any submodule, ensuring that I'm
using the same versions on all machines.

To install a new plugin:

    $ git submodule add https://github.com/user/plugin.git \
        pack/minpac/{start,opt}/plugin

Then add it to `packages.vim`.

To load an optional plugin use Vim's `:packadd` command.

To update installed plugins:

    :PackUpdate
    git add pack && git commit

Note that minpac isn't actually necessary, I'm just using it as an easy
way for me to update all submodules and review the new commits.

I did experiment with moving the existing submodules but noticed the
submodule names (which default to their path with in the parent repo)
weren't updated. It doesn't make sense for references to paths like
bundle/vimux to remain in my Git config files, and could cause confusion
in future.