add-plugin script for adding plugins as submodules
I'm not using minpac to install new plugins because whenever you use it
to update your plugins you get whatever's on the tip of the branch. That
will mean I'll end up running different versions on different machines,
which could be seriously painful. I want to use git to version the
plugins I'm running.

minpac is useful for updating submodules, and for skimming their
changelogs.

So to install a plugin I simply clone it into the appropriate directory
with `git submodule add`.

Moving a plugin from `pack/minpac/start` to `opt` gets a bit more
complicated; we need to:

- Remove the submodule from the `pack/minpac/start` directory
- Add the submodule within the `pack/minpac/opt` directory
- Update the call to `minpac#add()` in packages.vim

This script automates the first two steps.