Don't highlight matching parentheses for long
When you type a closing bracket the cursor jumps back to the opening
bracket, to highlight which set of parentheses I've just closed. By
default it lingers on the opening bracket for half a second, which I've
always found too long and a bit jarring.

Setting `matchtime` to 2 reduces it to 0.2 seconds.