Add shellcheck plugin
To use it, run :ShellCheck in a buffer containing a shell script. When
navigating the errors it outputs in the quickfix list you can use `gb`
to see the explanation on GitHub.

You see the raw shellcheck output before the quickfix list is populated
if you run shellcheck as a compiler:

    :compiler shellcheck
    :make