Add fzf and fzf.vim plugins
fzf.vim depends on the fzf plugin. I've installed the fzf and bat
command line tools separately.

Damn this is good! Commands to checkout include:

    :Files
    :GFiles     # search the files managed by git
    :Lines      # search through lines in open buffers
    :Blines     # search lines in current buffer
    :Buffers
    :Tags       # search ctags (also BTags for within current buffer)
    :Maps       # search your Vim mappings!
    :Colors

There's some awesome Git support too (that depends on fugitive).

    :Commits
    :BCommits   # select lines with visual mode to restrict search!

Note that `:Files` takes an optional path to search.

Use the bang version of a command to open in fullscreen mode (e.g. `:Files!`).

See [1] and [2] for more...

[1] https://github.com/junegunn/fzf.vim
[2] https://www.redhat.com/sysadmin/five-vim-plugins