Set autoread option
When an open buffer's file changes on disk, Vim warns you the next time
you try and change anything in the buffer. I always want it to refresh
the file from disk first, and setting `autoread` will enable that.

It doesn't always notice though; it's not polling. Run `:checkt[ime]` to
force Vim to check the timestamp on the files that back all open buffers.