Remove YankRang plugin
I found it was somehow interfering with my mapping for opening fuzzy
finders in Neovim, which I have set to <C-p>. Sometimes it would work,
sometimes it wouldn't.

I've never used the YankRing, and it appears that interfering with
mappings is a more widespread problem. [1]

I should try using registers as an alternative.

[1] https://www.reddit.com/r/vim/comments/4u5d0n/alternative_to_yankring/