Use .git/tags files
This setting was previously taken care of by the vim-fugitive plugin,
but was disabled by Tim Pope back in 2019. [1]

He explains his rationale in comments on the commit.

Basically, he's now suggesting that it's more sensible to just generate
your tags in ./tags at the root of your project, and then add the file
to `.gitignore`.

Keeping the feature in Fugitive also comes with a maintenance cost, so
he's deleted it.

Fair comment, but all my projects are setup to use `.git/tags` and right
now I want a quick fix to get my tag navigation working again.

[1] https://github.com/tpope/vim-fugitive/commit/63a05a69