Support Python's blue code formatter
Black is great and all that, but it doesn't quite conform to the Python
community's long standing conventions. In particular, Black has picked
the wrong kind of quote and overdone it on the line length.

So I'll be using Blue where possible instead.

https://pypi.org/project/blue/

The code in this `autocmd` was borrowed from a discussion on
StackExchange [1].

Note the use the `silent write` command. This prevents Vim from
complaining that the file has changed since it was last written (which
is true, if `blue` just modified it in place). We call `edit!` to reload
the file after `blue` has modified it.

The `silent write` hack/tweak is courtesy of [2].

[1] https://vi.stackexchange.com/questions/21507/is-it-possible-to-have-vim-run-a-command-every-time-save-w-is-invoked
[2] https://stackoverflow.com/questions/72127305/prevent-vim-warning-after-autocommand-changes-file-being-edited